Changes on the registration and login pages + added jest for unit tests

+++ .eslintrc.js
@@ -19,6 +19,6 @@ module.exports = {
         "react"
     ],
     "rules": {
-        "quotes": ["error", "single", { "avoidEscape": true, "allowTemplateLiterals": true }]
+        "quotes": ["error", "double", { "avoidEscape": true, "allowTemplateLiterals": true }]
     }
 };
